Registered 'undeployService' operation properties and some cleanup
o Registered 'undeployService' operation properties for Deploy to wire through the unDeploy service process.
o Removed unused parameter 'additionalAttributes' from container mapping configuration. The backing parameter will left in place to not cause unnecessary failures. But the parameter was and is not used so there is no functional impact. This fix cleans up the UI by not displaying an unused parameter.
o Added 'Undeploy Procedure' to the help file - the procedure itself was added in a previous release.